+++ docker/build.sh
@@ -5,8 +5,9 @@ set -o pipefail
 
 SDK_VERSION=1.0.0
 QT_VERSION_MINOR=5.15
-QT_VERSION_PATCH=8
+QT_VERSION_PATCH=19
 QT_VERSION=$QT_VERSION_MINOR.$QT_VERSION_PATCH
+IMAGE=unfoldedcircle/r2-toolchain-qt-$QT_VERSION-static
 
 VERSION_ARGS="\
 --build-arg BUILDROOT_SDK_VERSION=$SDK_VERSION \
@@ -14,11 +15,11 @@ VERSION_ARGS="\
 --build-arg QT_VERSION_PATCH=$QT_VERSION_PATCH"
 
 # use a local webserver to speed up archive downloads
-# Qt mirrors: https://download.qt.io/archive/qt/5.15/5.15.8/single/qt-everywhere-opensource-src-5.15.8.tar.xz.mirrorlist
+# Qt mirrors: https://download.qt.io/archive/qt/5.15/5.15.19/single/qt-everywhere-opensource-src-5.15.19.tar.xz.mirrorlist
 # QT_MIRROR URL must contain base path **before** `/archive/qt/`
-# Local file server example for SDK_VERSION=1.0.0, QT_VERSION_MINOR=5.15, QT_VERSION_PATCH=8 with files located at:
+# Local file server example for SDK_VERSION=1.0.0, QT_VERSION_MINOR=5.15, QT_VERSION_PATCH=19 with files located at:
 # - toolchain: /archive/ucr2/v1.0.0/ucr2-aarch64-toolchain-1.0.0-noqt.tar.gz
-# - qt:        /qt/5.15/5.15.8/single/qt-everywhere-opensource-src-5.15.8.tar.xz
+# - qt:        /qt/5.15/5.15.19/single/qt-everywhere-opensource-src-5.15.19.tar.xz
 #MIRROR_ARGS="\
 #--build-arg SDK_BASE_URL=http://172.16.16.10/archive/ucr2 \
 #--build-arg QT_MIRROR=http://172.16.16.10"
@@ -29,5 +30,5 @@ BUILD_LABELS="\
 --build-arg REVISION=$(git log -1 --format="%H")"
 
 docker build $VERSION_ARGS $MIRROR_ARGS $BUILD_LABELS \
-    -t unfoldedcircle/r2-toolchain-qt-$QT_VERSION-static \
-    -t unfoldedcircle/r2-toolchain-qt-$QT_VERSION-static:$SDK_VERSION .
+    -t $IMAGE \
+    -t $IMAGE:$SDK_VERSION .